server/requirements.txt
openai==1.12.0
langchain==0.1.6
langchain-community==0.0.19
langchain-openai==0.0.6
PyMuPDF==1.23.26
chromadb==0.4.22
tiktoken==0.5.2
flask==3.0.0
flask-cors==4.0.0
beautifulsoup4==4.12.3
requests==2.31.0
python-dotenv==1.0.0
Pillow==10.2.0